requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
classy-vision==0.6.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
Cython==0.29.30
easydict==1.9
einops==0.3.0
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
huggingface-hub==0.0.12
hydra-core==1.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iopath==0.1.9
joblib==1.1.0
lightning-bolts==0.3.4
llvmlite==0.38.1
Markdown==3.3.7
multidict==6.0.2
ndjson==0.3.1
numba==0.55.2
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.1.2
packaging==21.3
Pillow==9.1.1
portalocker==2.4.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
Pygments==2.12.0
pyparsing==3.0.9
pytorch-lightning==1.3.8
PyYAML==5.4.1
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rich==10.4.0
rsa==4.8
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torchmetrics==0.7.3
tqdm==4.46.0
transformers==4.9.1
tslearn==0.5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.9
Werkzeug==2.1.2
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
